requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
flask_mail
pycrypto
bcrypt
